requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.1
click==8.1.7
cytoolz==0.12.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.2.1
eth-utils==4.1.0
eth_abi==5.1.0
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
parsimonious==0.10.0
protobuf==5.26.1
pycryptodome==3.20.0
Pygments==2.17.2
python-coinmarketcap==0.5
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pyunormalize==15.1.0
redis==5.0.3
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rlp==4.0.0
rpds-py==0.18.0
rq==1.16.1
sniffio==1.3.1
toolz==0.12.1
typing_extensions==4.11.0
uniswap-python==0.7.1
urllib3==2.2.1
web3==6.17.2
websockets==12.0
yarl==1.9.4
